deployspecs
