+++ package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "start:dev": "webpack -w & nodemon server/app.js",
+    "start:dev": "webpack -w & nodemon server/index.js",
     "start": "npm run start:dev",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
